Put tax return signature routes behind portal authentication
