+++ app/views/portal/tax_returns/authorize_signature.html.erb
@@ -3,7 +3,7 @@
 <div class="slab consent-form">
   <%= render 'authorize_signature_header' %>
 
-  <%= form_for @form, url: tax_return_sign_path, method: :put, local: true, builder: VitaMinFormBuilder do |f| %>
+  <%= form_for @form, url: portal_tax_return_sign_path, method: :put, local: true, builder: VitaMinFormBuilder do |f| %>
     <div class="consent-checkboxes">
       <%= f.cfa_checkbox(:primary_accepts_terms, t(".efile_acceptance", year: @tax_return.year), options: { checked_value: "yes", unchecked_value: "no", classes: ['checkbox--full-width'] }) %>
       <%= f.cfa_checkbox(:primary_confirms_identity, t(".efile_identity_primary", name: @tax_return.client.legal_name.upcase, year: @tax_return.year), options: { checked_value: "yes", unchecked_value: "no", classes: ['checkbox--full-width'] }) %>
