+++ config/routes.rb
@@ -47,14 +47,6 @@ Rails.application.routes.draw do
       end
     end
 
-    resources :tax_returns, only: [] do
-      get '/sign', to: 'tax_returns#authorize_signature', as: :authorize_signature
-      put '/sign', to: 'tax_returns#sign', as: :sign
-      get '/spouse_sign', to: 'tax_returns#spouse_authorize_signature', as: :spouse_authorize_signature
-      put '/spouse_sign', to: 'tax_returns#spouse_sign', as: :spouse_sign
-      get '/success', to: 'tax_returns#success', as: :success
-    end
-
     resources :documents, only: [:destroy], controller: :documents do
       collection do
         DocumentNavigation.controllers.uniq.each do |controller_class|
@@ -123,6 +115,13 @@ Rails.application.routes.draw do
         get "locked", to: "client_logins#account_locked", as: :account_locked, on: :collection
         get "link-sent", to: "client_logins#link_sent", as: :login_link_sent, on: :collection
         get "invalid-token", to: "client_logins#invalid_token", as: :invalid_token, on: :collection
+      end
+      resources :tax_returns, only: [], path: '/tax-returns' do
+        get '/sign', to: 'tax_returns#authorize_signature', as: :authorize_signature
+        put '/sign', to: 'tax_returns#sign', as: :sign
+        get '/spouse-sign', to: 'tax_returns#spouse_authorize_signature', as: :spouse_authorize_signature
+        put '/spouse-sign', to: 'tax_returns#spouse_sign', as: :spouse_sign
+        get '/success', to: 'tax_returns#success', as: :success
       end
     end
