+++ db/schema.rb
@@ -10,7 +10,6 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-
 ActiveRecord::Schema.define(version: 2021_01_15_165125) do
 
   # These are extensions that must be enabled in order to support this database
